+++ src/KMonad/Args/Parser.hs
@@ -290,6 +290,7 @@ keywordButtons =
   , ("layer-rem"      , KLayerRem    <$> lexeme word)
   , ("layer-delay"    , KLayerDelay  <$> lexeme numP <*> lexeme word)
   , ("layer-next"     , KLayerNext   <$> lexeme word)
+  , ("hold-button"    , pure KHoldButton)
   , ("around-next"    , KAroundNext  <$> buttonP)
   , ("around-next-single", KAroundNextSingle <$> buttonP)
   , ("before-after-next", KBeforeAfterNext <$> buttonP <*> buttonP)
@@ -301,6 +302,7 @@ keywordButtons =
   , ("cmd-button"     , KCommand     <$> lexeme textP <*> optional (lexeme textP))
   , ("pause"          , KPause . fromIntegral <$> lexeme numP)
   , ("sticky-key"     , KStickyKey   <$> lexeme numP <*> buttonP)
+  , ("retap"          , KRetap       <$> lexeme numP <*> buttonP <*> buttonP)
   ]
   ++ map (\(nm,_,btn) -> (nm, btn <$> buttonP <*> buttonP)) implArndButtons
  where